shovel "done"
Changed the texture to something a little less ridiculous. Also implemented the functionality for toggling the 3x3 destruction mode, which can now be implemented for the pickaxe as well. Only problem is that when in 3x3 mode it insta destroys all the blocks, which id like to try to fix.