Unkempt Harold projectile
Unkempt Harold now shoots an exploding projectile with a radius of 5 and the damage to mobs from it has been capped at 13.5, which is the highest possible hit for it otherwise. The texture and potentially model for the projectile still needs to be made and the animation for the shovel should be made as well.